backend/pyproject.toml
[tool.poetry]
name = "sns-solution-backend"
version = "0.1.0"
description = "SNS Operation Management Integrated Solution - Backend"
authors = ["SNS Solution Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
alembic = "^1.14.0"
redis = {extras = ["hiredis"], version = "^5.2.0"}
celery = {extras = ["redis"], version = "^5.4.0"}
pydantic-settings = "^2.6.0"
pydantic = {extras = ["email"], version = "^2.10.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
httpx = "^0.28.0"
python-magic-bin = "^0.4.14"
sentry-sdk = {extras = ["fastapi", "celery"], version = "^2.19.0"}
structlog = "^24.4.0"
python-multipart = "^0.0.18"
boto3 = "^1.35.0"
cryptography = "^44.0.0"
bcrypt = "4.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
httpx = "^0.28.0"
ruff = "^0.8.0"
mypy = "^1.13.0"
factory-boy = "^3.3.0"
aiosqlite = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]
ignore = ["A003"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
